# api-spec-json.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Alpaca API Service",
    "description": "A lightweight API service for Alpaca Markets integration with artificial order management and centralized configuration",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/yourusername/alpaca-service"
    }
  },
  "servers": [
    {
      "url": "http://localhost:9000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/orders": {
      "post": {
        "summary": "Place order",
        "tags": ["Orders"],
        "operationId": "placeOrder",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order placed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit or cooldown",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get orders",
        "tags": ["Orders"],
        "operationId": "getOrders",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["open", "closed", "all"]
            }
          },
          {
            "name": "symbol",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "side",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["buy", "sell"]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 50
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of orders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OrderRequest": {
        "type": "object",
        "required": ["symbol", "qty", "side", "orderType", "timeInForce"],
        "properties": {
          "symbol": {
            "type": "string",
            "description": "Stock symbol"
          },
          "qty": {
            "type": "number",
            "description": "Order quantity"
          },
          "side": {
            "type": "string",
            "enum": ["buy", "sell"],
            "description": "Order side"
          },
          "orderType": {
            "type": "string",
            "enum": ["market", "limit", "stop", "stop_limit"],
            "description": "Order type"
          },
          "timeInForce": {
            "type": "string",
            "enum": ["day", "gtc", "opg", "cls", "ioc", "fok"],
            "description": "Time in force"
          },
          "limitPrice": {
            "type": "number",
            "description": "Limit price (required for limit orders)"
          },
          "stopPrice": {
            "type": "number",
            "description": "Stop price (required for stop orders)"
          },
          "clientOrderId": {
            "type": "string",
            "description": "Client-specified order ID"
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "orderId": { "type": "string" },
          "clientOrderId": { "type": "string" },
          "symbol": { "type": "string" },
          "qty": { "type": "number" },
          "filledQty": { "type": "number" },
          "side": { "type": "string" },
          "orderType": { "type": "string", "readOnly": true, "description": "Order type derived by the service based on provided parameters" },
          "timeInForce": { "type": "string" },
          "limitPrice": { "type": "number" },
          "stopPrice": { "type": "number" },
          "status": { "type": "string" },
          "filledAvgPrice": { "type": "number" },
          "duplicateCheckPassed": { "type": "boolean" },
          "rateLimitStatus": { "type": "string" },
          "notes": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "submittedAt": { "type": "string", "format": "date-time" },
          "filledAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["error"]
          },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "detail": { "type": "string" },
              "fields": { "type": "object", "description": "Map of field names to error messages for validation errors" },
              "metadata": { "type": "object" }
            }
          }
        }
      },
...
      "OrderResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "data": { "$ref": "#/components/schemas/Order" }
            }
          }
        ]
      },
      "OrdersResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "orders": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/Order" }
                  },
                  "nextCursor": { "type": "string" }
                }
              }
            }
          }
        ]
      },
...
    }
  },
...
}